+++ tests/testthat/tests_tidier.R
@@ -3,12 +3,12 @@ test_that("basic mutate", {
   res = iris %>%
     mutate(sl_pl_1 = Sepal.Length + 1)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = iris %>%
     mutate_(sl_pl_1 = Sepal.Length + 1)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 })
 
 test_that("order_by without by", {
@@ -19,7 +19,7 @@ test_that("order_by without by", {
            ) %>%
     dplyr::arrange(Petal.Width)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   # involved order_by
   res = iris %>%
@@ -28,7 +28,7 @@ test_that("order_by without by", {
             ) %>%
     dplyr::arrange(desc(Petal.Width), Sepal.Length)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = iris %>%
     mutate_(sl_pl_1 = cumsum(Sepal.Length),
@@ -36,7 +36,7 @@ test_that("order_by without by", {
             ) %>%
     dplyr::arrange(Petal.Width)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = iris %>%
     mutate_(sl_pl_1 = cumsum(Sepal.Length),
@@ -45,7 +45,7 @@ test_that("order_by without by", {
             ) %>%
     dplyr::arrange(desc(Petal.Width), Sepal.Length)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 })
 
 test_that("order_by with by", {
@@ -59,7 +59,7 @@ test_that("order_by with by", {
     dplyr::select(-rn) %>%
     dplyr::slice_min(n = 3, Petal.Width, by = Species)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = iris %>%
     dplyr::mutate(rn = dplyr::row_number()) %>%
@@ -70,7 +70,7 @@ test_that("order_by with by", {
     dplyr::select(-rn) %>%
     dplyr::slice_min(n = 3, Petal.Width, by = Species)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 })
 
 test_that("order_by, with by, with frame", {
@@ -83,7 +83,7 @@ test_that("order_by, with by, with frame", {
             ) %>%
     dplyr::slice_min(n = 3, Petal.Width, by = Species)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = iris %>%
     mutate_(sl_mean = mean(Sepal.Length),
@@ -93,7 +93,7 @@ test_that("order_by, with by, with frame", {
             ) %>%
     dplyr::slice_min(n = 3, Petal.Width, by = Species)
 
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 })
 
 test_that("order_by, with by, with frame, with index", {
@@ -126,7 +126,7 @@ test_that("order_by, with by, with frame, with index", {
                     ),
          .index = date_col
          )
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 
   res = airquality %>%
   # create date column
@@ -156,5 +156,114 @@ test_that("order_by, with by, with frame, with index", {
                     ),
          .index = "date_col"
          )
-  testthat::expect(inherits(res, "data.frame"))
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+})
+
+test_that("order_by, with by, with frame, same column name", {
+
+  res = iris %>%
+    mutate(Sepal.Length = mean(Sepal.Length),
+            .frame = c(Inf, 0),
+            .order_by = Petal.Width,
+            .by = c(Petal.Length, Sepal.Width)
+            ) %>%
+    dplyr::slice_min(n = 3, Petal.Width, by = Species)
+
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+
+  res = iris %>%
+    mutate_(Sepal.Length = mean(Sepal.Length),
+            .frame = c(Inf, 0),
+            .order_by = "Petal.Width",
+            .by = c("Species", "Petal.Length")
+            ) %>%
+    dplyr::slice_min(n = 3, Petal.Width, by = Species)
+
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+})
+
+test_that("order_by, with by, with frame, same column name", {
+
+  res = iris %>%
+    mutate(Sepal.Length = mean(Sepal.Length),
+            .frame = c(Inf, 0),
+            .order_by = Petal.Width,
+            .by = c(Petal.Length, Sepal.Width)
+            ) %>%
+    dplyr::slice_min(n = 3, Petal.Width, by = Species)
+
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+
+  res = iris %>%
+    mutate_(Sepal.Length = mean(Sepal.Length),
+            .frame = c(Inf, 0),
+            .order_by = "Petal.Width",
+            .by = c("Species", "Petal.Length")
+            ) %>%
+    dplyr::slice_min(n = 3, Petal.Width, by = Species)
+
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+})
+
+test_that("order_by, with by, with frame, with index, with same column", {
+
+  res = airquality %>%
+  # create date column
+  dplyr::mutate(date_col = as.Date(paste("1973",
+                                         stringr::str_pad(Month,
+                                                          width = 2,
+                                                          side = "left",
+                                                          pad = "0"
+                                                          ),
+                                         stringr::str_pad(Day,
+                                                          width = 2,
+                                                          side = "left",
+                                                          pad = "0"
+                                                          ),
+                                         sep = "-"
+                                         )
+                                  )
+                ) %>%
+  # create gaps by removing some days
+  dplyr::slice_sample(prop = 0.8) %>%
+  # compute mean temperature over last seven days in the same month
+  mutate(Temp = mean(Temp, na.rm = TRUE),
+         .order_by = Day,
+         .by = Month,
+         .frame = c(lubridate::days(7), # 7 days before current row
+                    lubridate::days(-1) # do not include current row
+                    ),
+         .index = date_col
+         )
+  testthat::expect(inherits(res, "data.frame"), "not a df")
+
+  res = airquality %>%
+  # create date column
+  dplyr::mutate(date_col = as.Date(paste("1973",
+                                         stringr::str_pad(Month,
+                                                          width = 2,
+                                                          side = "left",
+                                                          pad = "0"
+                                                          ),
+                                         stringr::str_pad(Day,
+                                                          width = 2,
+                                                          side = "left",
+                                                          pad = "0"
+                                                          ),
+                                         sep = "-"
+                                         )
+                                  )
+                ) %>%
+  # create gaps by removing some days
+  dplyr::slice_sample(prop = 0.8) %>%
+  # compute mean temperature over last seven days in the same month
+  mutate_(Temp = mean(Temp, na.rm = TRUE),
+         .order_by = "Day",
+         .by = "Month",
+         .frame = c(lubridate::days(7), # 7 days before current row
+                    lubridate::days(-1) # do not include current row
+                    ),
+         .index = "date_col"
+         )
+  testthat::expect(inherits(res, "data.frame"), "not a df")
 })