+++ NEWS.md
@@ -1,3 +1,8 @@
+# tidier 0.1.0 (on github: 2023-06-01)
+
+* Exposed slider's `.complete` argument in `tidier::mutate`
+* bugfix: `mutate` can now modify a column (same name) in sliding operation. 
+
 # tidier 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.